requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.0
charset-normalizer==3.4.1
cryptography==43.0.0
distlib==0.3.8
dj-rest-auth==7.0.1
Django==5.1.6
django-allauth==0.63.6
django-environ==0.10.0
django-extensions==3.2.3
djangorestframework==3.14.0
drf-yasg==1.21.7  # Swagger/OpenAPI docs
filelock==3.15.4
gunicorn==20.1.0
idna==3.10
MarkupSafe==2.1.5
mssql-django==1.1
mysqlclient==2.2.4
pillow==10.4.0
platformdirs==4.2.2
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
pyodbc==5.1.0
pyOpenSSL==24.2.1
pytz==2024.1
requests==2.32.3
setuptools==72.1.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.3.0
virtualenv==20.26.3
Werkzeug==3.0.3
whitenoise==6.2.0
coreapi==2.3.3
openapi-codec==1.3.2